// src/types/api.ts
export interface Character {
  id: number;
  name: string;
  status: 'Alive' | 'Dead' | 'unknown';
  species: string;
  type: string;
  gender: 'Female' | 'Male' | 'Genderless' | 'unknown';
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface ApiResponse<T> {
  info: {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
  };
  results: T[];
}

export interface CharacterFilters {
  status?: 'Alive' | 'Dead' | 'unknown';
  gender?: 'Female' | 'Male' | 'Genderless' | 'unknown';
  page?: number;
}

export interface CharacterQueryParams {
  status?: string;
  gender?: string;
  page?: string;
}